# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestion des Recettes",
        "description": "API pour gérer les utilisateurs, recettes, catégories, ingrédients et favoris.",
        "contact": {
            "email": "support@votreapp.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/ingredients": {
            "get": {
                "tags": [
                    "Ingredients"
                ],
                "summary": "Get a list of all ingredients",
                "description": "Returns a list of all ingredients in the system.",
                "operationId": "39ca45df16ed77cc6fa4231d4b0829ba",
                "responses": {
                    "200": {
                        "description": "A list of ingredients",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Ingredient"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Ingredients"
                ],
                "summary": "Create a new ingredient",
                "description": "Creates a new ingredient with the provided data.",
                "operationId": "52194d093351ba6fd994349231ed08be",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Sugar"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ingredient created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ingredient"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/ingredients/{id}": {
            "get": {
                "tags": [
                    "Ingredients"
                ],
                "summary": "Get ingredient by ID",
                "description": "Returns an ingredient by its ID.",
                "operationId": "e1429941b49f51784908478c60e9d0df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the ingredient",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ingredient data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ingredient"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ingredient not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ingredient not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Ingredients"
                ],
                "summary": "Update ingredient by ID",
                "description": "Updates an ingredient's details by its ID.",
                "operationId": "78e31c455931754bfabbd087abb78e78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the ingredient to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Brown Sugar"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ingredient updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ingredient"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ingredient not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ingredient not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Ingredients"
                ],
                "summary": "Delete ingredient by ID",
                "description": "Deletes an ingredient by its ID.",
                "operationId": "c9be8d333be55722b7962cfec274c556",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the ingredient to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Ingredient deleted successfully"
                    },
                    "404": {
                        "description": "Ingredient not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ingredient not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Ingredient": {
                "title": "Ingredient",
                "description": "Ingredient model",
                "properties": {
                    "": {
                        "title": "Recipes",
                        "description": "List of recipes that include this ingredient",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Recipe"
                        }
                    },
                    "id": {
                        "title": "ID",
                        "description": "ID of the ingredient",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "fillable": {
                        "title": "Name",
                        "description": "Name of the ingredient",
                        "type": "string",
                        "example": "Sugar"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Ingredient"
                }
            },
            "Recipe": {
                "title": "Recipe",
                "description": "Recipe model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID of the recipe",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "title": "Title",
                        "description": "Title of the recipe",
                        "type": "string",
                        "example": "Delicious Pancakes"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Brief description of the recipe",
                        "type": "string",
                        "example": "A simple and quick pancake recipe."
                    },
                    "instructions": {
                        "title": "Instructions",
                        "description": "Detailed cooking instructions for the recipe",
                        "type": "string",
                        "example": "1. Mix ingredients... 2. Cook on medium heat..."
                    },
                    "user_id": {
                        "title": "User ID",
                        "description": "ID of the user who created the recipe",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "category_id": {
                        "title": "Category ID",
                        "description": "ID of the category this recipe belongs to",
                        "type": "integer",
                        "format": "int64",
                        "example": 2
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Recipe"
                }
            },
...
        }
    },
...
}